requirements.txt
fastapi==0.100.1
uvicorn[standard]==0.22.0
httpx==0.24.1
python-dotenv==1.1.0
